Cliente front-back y usuario

+++ src/main/resources/static/js/funciones-genericas.js
@@ -1,5 +1,5 @@
 function loadContent(page) {
-    $('#contenido-principal').load(page, function(response, status, xhr) {
+    $('#contenido-principal').load(page, function (response, status, xhr) {
         if (status === "error") {
             var msg = "Lo sentimos, hubo un error: ";
             $("#contenido-principal").html(msg + xhr.status + " " + xhr.statusText);
@@ -9,14 +9,14 @@ function loadContent(page) {
 
 
 // Función para crear y mostrar el modal
-function showAlertModal(message, type = 'success', automodal) {
+function showAlertModal(message, type = 'success', title = "Alerta", automodal) {
     // Crear el modal HTML
     const modalHTML = `
     <div class="modal fade" id="alertModalGenerica" tabindex="-1" aria-labelledby="alertModalLabel" aria-hidden="true">
       <div class="modal-dialog" role="document">
         <div class="modal-content">
           <div class="modal-header">
-            <h5 class="modal-title" id="alertModalLabel">Alert</h5>
+            <h5 class="modal-title" id="alertModalLabel">${title}</h5>
             <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
           </div>
           <div class="modal-body">
@@ -51,8 +51,8 @@ function showAlertModal(message, type = 'success', automodal) {
             }
         });
 
-        if(automodal) {
-            setTimeout(function() {
+        if (automodal) {
+            setTimeout(function () {
                 restoreBodyStyles(modalInstance, modalElement);
             }, 2000);
         }
@@ -64,15 +64,50 @@ function showAlertModal(message, type = 'success', automodal) {
 function restoreBodyStyles(modalInstance, modalElement) {
     const backdrop = document.querySelector('.modal-backdrop');
     if (backdrop) {
-            modalInstance.dispose();
-            modalElement.remove();
-            document.body.classList.remove('modal-open'); // Elimina la clase modal-open del body
-            backdrop.remove();
-            document.body.style.overflow = ''; // Restaurar el estilo de overflow
+        modalInstance.dispose();
+        modalElement.remove();
+        document.body.classList.remove('modal-open'); // Elimina la clase modal-open del body
+        backdrop.remove();
+        document.body.style.overflow = ''; // Restaurar el estilo de overflow
     }
 }
 
-document.getElementById("logout").addEventListener("click", () => {
-    localStorage.clear();
-    window.location.href = "/app/index.html";
-});
+if (document.getElementById("logout")) {
+    document.getElementById("logout").addEventListener("click", () => {
+        localStorage.clear();
+        window.location.href = "/app/index.html";
+    });
+}
+
+if (document.getElementById("titulo-usuario")) {
+    document.getElementById("titulo-usuario").innerText = localStorage.getItem("usuario");
+}
+
+
+function editableTDNumber(td) {
+    td.textContent = td.textContent.replace(/[^0-9]/g, '');
+}
+
+function testTDNumber(event) {
+    const key = event.key;
+    if (!/^[0-9\b]|Backspace|Delete|Tab|ArrowLeft|ArrowRight|ArrowUp|ArrowDown/.test(key)) {
+        event.preventDefault();
+    }
+}
+
+function comprobarCamposObligatorios(nombre_formulario) {
+
+    let formulario = document.querySelectorAll("#" + nombre_formulario + " input, " + "#" + nombre_formulario + " select");
+
+    let vacios = Array.from(formulario)
+        .filter(fld => fld.required && (fld.tagName === 'SELECT' ? fld.value === '' : fld.value.trim() === ''))
+        .map(fld => "<b>" + (fld.placeholder || fld.ariaLabel) + "</b>"); // Usa el placeholder o el nombre del campo
+
+    return vacios;
+}
+
+function evitarTeclas(campo, event) {
+    campo.value="";
+    event.preventDefault();
+}
+
